requirements.txt
# Core ML Frameworks
torch>=1.9.0
torchvision>=0.10.0
tensorflow>=2.8.0
transformers>=4.20.0

# Vision-Language Models
accelerate>=0.20.0
bitsandbytes>=0.40.0
einops>=0.6.0
safetensors>=0.3.0
peft>=0.4.0
transformers-stream-generator>=0.0.4

# Computer Vision
opencv-python>=4.5.0
Pillow>=8.3.0
efficientnet-pytorch>=0.7.1

# Data Science
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0

# Evaluation Metrics
editdistance>=1.0.0
pycocoevalcap>=1.0.0
nltk>=3.7.0

# Visualization & Utilities
matplotlib>=3.4.0
tqdm>=4.60.0
jupyter>=1.0.0

# Additional Dependencies
timm>=0.6.0
datasets>=2.0.0
huggingface-hub>=0.15.0